Couple more records ripped
This is the 2nd commit message:

I think I finally got the imports right (relative imports to the rescue from
bash/bash)

---------------------- Mopy/bash/game/oblivion/records.py ----------------------
 from . import esp

 #--Mod I/O
-from bash.bolt import StateError, Flags
-from bash.brec import MelRecord
+from ...bolt import StateError, Flags
+from ...brec import MelRecord, BaseRecordHeader, ModError

-class RecordHeader(bash.brec.BaseRecordHeader):
+class RecordHeader(BaseRecordHeader):
     size = 20

     def __init__(self,recType='TES4',size=0,arg1=0,arg2=0,arg3=0,*extra):
@@ -54,7 +54,7 @@ def unpack(ins):
         type,size,uint0,uint1,uint2 = ins.unpack('=4s4I',20,'REC_HEADER')
         #--Bad?
         if type not in esp.recordTypes:
-            raise bash.brec.ModError(ins.inName,u'Bad header type: '+repr(type))
+            raise ModError(ins.inName,u'Bad header type: '+repr(type))
         #--Record
         if type != 'GRUP':
             pass
@@ -66,7 +66,7 @@ def unpack(ins):
             elif str0 in esp.topIgTypes:
                 uint0 = esp.topIgTypes[str0]
             else:
-                raise bash.brec.ModError(ins.inName,u'Bad Top GRUP type: '+repr(str0))
+                raise ModError(ins.inName,u'Bad Top GRUP type: '+repr(str0))
         return RecordHeader(type,size,uint0,uint1,uint2)

     def pack(self):
@@ -100,7 +100,7 @@ class MelBipedFlags(bash.bolt.Flags):
     """Biped flags element. Includes biped flag set by default."""
     mask = 0xFFFF
     def __init__(self,default=0L,newNames=None):
-        names = bolt.Flags.getNames('head', 'hair', 'upperBody', 'lowerBody', 'hand', 'foot', 'rightRing', 'leftRing', 'amulet', 'weapon', 'backWeapon', 'sideWeapon', 'quiver', 'shield', 'torch', 'tail')
+        names = Flags.getNames('head', 'hair', 'upperBody', 'lowerBody', 'hand', 'foot', 'rightRing', 'leftRing', 'amulet', 'weapon', 'backWeapon', 'sideWeapon', 'quiver', 'shield', 'torch', 'tail')
         if newNames: names.update(newNames)
         Flags.__init__(self,default,names)
